Move updates into own file
